+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jfelty <jfelty@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/23 15:24:48 by jfelty            #+#    #+#             */
-/*   Updated: 2019/10/25 14:30:37 by jfelty           ###   ########.fr       */
+/*   Updated: 2019/10/29 18:29:44 by jfelty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -102,7 +102,7 @@ int					ft_power(int nb, int power);
 char				*ft_strrev(char *str);
 char				*ft_ll_itoa_base(uint64_t n, int base, int return_upper);
 char				*ft_push_to_str_front(char *str, char c);
-int					ft_makepos(int a);
+int					ft_makepos(int64_t a);
 int					ft_largest_int(int a, int b);
 int					ft_isneg(int64_t a);
 #endif
